ArticleTableViewCellを分離してフォルダ整理しました。
